# paylab-ui/openapi/paylab.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/accounts": {
      "post": {
        "operationId": "AccountsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "operationId": "AccountsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque `nextCursor` from the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/v1/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or another Merchant's resource (indistinguishable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/v1/accounts/{id}/balance": {
      "get": {
        "operationId": "AccountsController_balance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or another Merchant's resource (indistinguishable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/v1/accounts/{id}/entries": {
      "get": {
        "operationId": "AccountsController_entries",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque `nextCursor` from the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerEntryPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or another Merchant's resource (indistinguishable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/v1/payments": {
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      },
      "get": {
        "operationId": "PaymentsController_list",
        "parameters": [
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Exclusive. Same format as `from`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive. ISO 8601 timestamp with a zone, or a YYYY-MM-DD day (00:00 UTC).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "CREATED",
                "PROCESSING",
                "SUCCEEDED",
                "FAILED"
              ],
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque `nextCursor` from the previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/v1/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown id or another Merchant's resource (indistinguishable).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/v1/reports/daily": {
      "get": {
        "operationId": "ReportsController_daily",
        "parameters": [
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Last UTC day, inclusive, YYYY-MM-DD. The range is at most 366 days.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "description": "First UTC day, inclusive, YYYY-MM-DD.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_handle",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    }
  },
  "info": {
    "title": "PayLab API",
    "description": "Payment processing and double-entry ledger lab.",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "WalletListItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "example": "WALLET"
          },
          "currency": {
            "type": "string",
            "example": "BRL"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "kind",
          "currency",
          "createdAt"
        ]
      },
      "WalletPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WalletListItemResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pass as `cursor` to get the following page; null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Invalid API key"
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "description": "Field-level details when available."
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
      "AccountResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "example": "WALLET"
          },
          "currency": {
            "type": "string",
            "example": "BRL"
          }
        },
        "required": [
          "id",
          "kind",
          "currency"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "RESOURCE_NOT_FOUND"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "balance": {
            "type": "number",
            "description": "Credits minus debits. Integer centavos, never a decimal."
          },
          "currency": {
            "type": "string",
            "example": "BRL"
          }
        },
        "required": [
          "accountId",
          "balance",
          "currency"
        ]
      },
      "LedgerEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ledgerTransactionId": {
            "type": "string",
            "format": "uuid"
          },
          "direction": {
            "type": "string",
            "enum": [
              "DEBIT",
              "CREDIT"
            ]
          },
          "amount": {
            "type": "number",
            "description": "Always positive; the sign comes from direction. Integer centavos, never a decimal."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "ledgerTransactionId",
          "direction",
          "amount",
          "createdAt"
        ]
      },
      "LedgerEntryPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerEntryResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pass as `cursor` to get the following page; null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "description": "Integer centavos, never a decimal."
          },
          "currency": {
            "type": "string",
            "example": "BRL"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "failureReason": {
            "type": "string",
            "nullable": true,
            "example": "INSUFFICIENT_FUNDS"
          },
          "ledgerTransactionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sourceAccountId",
          "destinationAccountId",
          "amount",
          "currency",
          "status",
          "failureReason",
          "ledgerTransactionId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaymentPageResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pass as `cursor` to get the following page; null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "DailyReportRowResponse": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-09-01",
            "description": "UTC calendar day."
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PROCESSING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "count": {
            "type": "number"
          },
          "volume": {
            "type": "number",
            "description": "Sum of the Payments' amounts. Integer centavos, never a decimal."
          }
        },
        "required": [
          "date",
          "status",
          "count",
          "volume"
        ]
      },
      "DailyReportResponse": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "example": "2026-09-01"
          },
          "to": {
            "type": "string",
            "example": "2026-09-30"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyReportRowResponse"
            }
          }
        },
        "required": [
          "from",
          "to",
          "items"
        ]
      }
    }
  }
}